update the file name
