+++ weevent-client/src/test/java/com/webank/weevent/client/sdk/WeEventClientTest.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import javax.net.ssl.SSLContext;
 
 import com.webank.weevent.sdk.BrokerException;
+import com.webank.weevent.sdk.ErrorCode;
 import com.webank.weevent.sdk.SendResult;
 import com.webank.weevent.sdk.TopicInfo;
 import com.webank.weevent.sdk.TopicPage;
@@ -20,7 +21,10 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TestName;
 
+import static java.lang.Thread.sleep;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
 /**
@@ -31,38 +35,25 @@ import static org.junit.Assert.assertTrue;
  * @since <pre>05/10/2019</pre>
  */
 @Slf4j
-public class WeEventClientTest {
-    private Map<String, String> extensions = new HashMap<>();
+public class WeEventClientRemoveGroupIdTest {
+
     @Rule
     public TestName testName = new TestName();
 
     public String topicName = "com.webank.weevent";
 
-    public String groupId = "1";
-
     private WeEventClient weEventClient;
 
     @Before
     public void before() throws Exception {
-        extensions.put("weevent-url", "https://github.com/WeBankFinTech/WeEvent");
         weEventClient = new WeEventClient("http://127.0.0.1:8080/weevent");
         weEventClient.open(topicName);
+
     }
 
     @After
     public void after() throws Exception {
         weEventClient.close(topicName);
-    }
-
-    /**
-     * Method: publish(String topic, String groupId, byte[] content, Map<String, String> extensions)
-     */
-    @Test
-    public void testPublishGroupId() throws Exception {
-        log.info("===================={}", this.testName.getMethodName());
-        // test groupId
-        SendResult sendResult = this.weEventClient.publish(topicName, groupId, "hello world".getBytes(StandardCharsets.UTF_8), extensions);
-        assertTrue(sendResult.getStatus() == SendResult.SendResultStatus.SUCCESS);
     }
 
 
@@ -72,51 +63,8 @@ public class WeEventClientTest {
     @Test
     public void testPublish() throws Exception {
         log.info("===================={}", this.testName.getMethodName());
-        SendResult sendResult = this.weEventClient.publish(topicName, "hello world".getBytes(StandardCharsets.UTF_8), extensions);
+        SendResult sendResult = this.weEventClient.publish(topicName, "hello world".getBytes(StandardCharsets.UTF_8));
         assertTrue(sendResult.getStatus() == SendResult.SendResultStatus.SUCCESS);
-    }
-
-    /**
-     * Method: publish(String topic, String groupId, byte[] content, Map<String, String> extensions)
-     */
-    @Test(expected = BrokerException.class)
-    public void testPublish_001() throws Exception {
-        log.info("===================={}", this.testName.getMethodName());
-
-        this.weEventClient.publish(this.topicName, groupId, null, null);
-    }
-
-    @Test(expected = BrokerException.class)
-    public void testPublish_002() throws Exception {
-        log.info("===================={}", this.testName.getMethodName());
-
-        this.weEventClient.publish(null, groupId, "hello world".getBytes(StandardCharsets.UTF_8), extensions);
-    }
-
-
-    @Test(expected = BrokerException.class)
-    public void testPublish_004() throws Exception {
-        log.info("===================={}", this.testName.getMethodName());
-        // test groupId
-        this.weEventClient.publish(topicName, groupId, "hello world".getBytes(StandardCharsets.UTF_8), extensions);
-    }
-
-    @Test(expected = BrokerException.class)
-    public void testPublish_005() throws Exception {
-        log.info("===================={}", this.testName.getMethodName());
-        // test groupId
-        this.weEventClient.publish(topicName, groupId, "hello world".getBytes(StandardCharsets.UTF_8), null);
-    }
-
-
-    /**
-     * Method: publish(String topic, byte[] content)
-     */
-    @Test(expected = BrokerException.class)
-    public void testPublish_003() throws Exception {
-        log.info("===================={}", this.testName.getMethodName());
-
-        this.weEventClient.publish("111111111111111111111111111111111111111111111111111111", null, null, extensions);
     }
 
 
@@ -127,10 +75,12 @@ public class WeEventClientTest {
     public void testSubscribe() throws Exception {
         log.info("===================={}", this.testName.getMethodName());
         // create subscriber
-        String subscriptionId = this.weEventClient.subscribe(this.topicName, WeEvent.OFFSET_LAST, new WeEventClient.EventListener() {
+        String groupId = "1";//if not set default 1
+        this.weEventClient.subscribe(this.topicName, groupId, WeEvent.OFFSET_LAST, new WeEventClient.EventListener() {
             @Override
             public void onEvent(WeEvent event) {
-                System.out.println(event.toString());
+                System.out.println("onEvent:" + event.toString());
+                log.info("onEvent:" + event.toString());
             }
 
             @Override
@@ -138,21 +88,36 @@ public class WeEventClientTest {
                 e.printStackTrace();
             }
         });
-        Thread.sleep(60000);
+        sleep(1000000);
     }
 
     /**
-     * Method: subscribe(String topic, groupId, String offset, IConsumer.ConsumerListener listener)
+     * test topic length > 64
      */
     @Test
-    public void testSubscribe_01() throws Exception {
+    public void testOpen_topicOverMaxLen() {
+        try {
+            String topic = "topiclengthexceeding64-123456789012345678901234567890123456789012";
+            boolean result = weEventClient.open(topic);
+            assertNull(result);
+        } catch (BrokerException e) {
+            assertEquals(e.getCode(), ErrorCode.TOPIC_EXCEED_MAX_LENGTH.getCode());
+        }
+    }
+
+    /**
+     * Method: subscribe(String topic, String offset, IConsumer.ConsumerListener listener)
+     */
+    @Test
+    public void testSubscribeEventId() throws Exception {
         log.info("===================={}", this.testName.getMethodName());
-        // create subscriber
+        SendResult sendResult = this.weEventClient.publish(topicName, "hello world".getBytes(StandardCharsets.UTF_8));
+        System.out.print("sendResult getEventId" + sendResult.getEventId());
         String groupId = "1";//if not set default 1
-        String subscriptionId = this.weEventClient.subscribe(this.topicName, groupId, "317e7c4c-8-26", new WeEventClient.EventListener() {
+        this.weEventClient.subscribe(this.topicName, groupId, sendResult.getEventId(), new WeEventClient.EventListener() {
             @Override
             public void onEvent(WeEvent event) {
-                log.info(event.toString());
+                System.out.print(event.toString());
             }
 
             @Override
@@ -160,8 +125,9 @@ public class WeEventClientTest {
                 e.printStackTrace();
             }
         });
-        Thread.sleep(60000);
+        sleep(100000);
     }
+
 
     /**
      * Method: unSubscribe(String subscriptionId)
@@ -197,15 +163,6 @@ public class WeEventClientTest {
 
 
     /**
-     * Method: open(String topic,String groupId)
-     */
-    @Test
-    public void testOpenGroupId() throws Exception {
-        boolean result = this.weEventClient.open(this.topicName, this.groupId);
-        assertTrue(result);
-    }
-
-    /**
      * Method: close(String topic)
      */
     @Test
@@ -216,25 +173,6 @@ public class WeEventClientTest {
 
 
     /**
-     * Method: testCloseGroupId(String topic,String groupId)
-     */
-    @Test
-    public void testCloseGroupId() throws Exception {
-        boolean result = weEventClient.close(topicName, groupId);
-        assertTrue(result);
-    }
-
-
-    /**
-     * Method: exist(String topic)
-     */
-    @Test
-    public void testExistGroupId() throws Exception {
-        boolean result = this.weEventClient.exist(this.topicName, groupId);
-        assertTrue(result);
-    }
-
-    /**
      * Method: exist(String topic)
      */
     @Test
@@ -242,26 +180,6 @@ public class WeEventClientTest {
         boolean result = this.weEventClient.exist(this.topicName);
         assertTrue(result);
     }
-
-
-    /**
-     * Method: exist(String topic)
-     */
-    @Test
-    public void testExist_001() throws Exception {
-        boolean result = this.weEventClient.exist("not exist", groupId);
-        assertFalse(result);
-    }
-
-    /**
-     * Method: testListGroupId(Integer pageIndex, Integer pageSize)
-     */
-    @Test
-    public void testListGroupId() throws Exception {
-        TopicPage list = this.weEventClient.list(0, 10, groupId);
-        assertTrue(list.getTotal() > 0);
-    }
-
 
     /**
      * Method: list(Integer pageIndex, Integer pageSize)
@@ -276,27 +194,11 @@ public class WeEventClientTest {
      * Method: state(String topic)
      */
     @Test
-    public void testStateGroupId() throws Exception {
-        TopicInfo info = this.weEventClient.state(this.topicName, groupId);
-        assertTrue(info.getTopicName().equals(this.topicName));
-    }
-
-    /**
-     * Method: state(String topic)
-     */
-    @Test
     public void testState() throws Exception {
         TopicInfo info = this.weEventClient.state(this.topicName);
         assertTrue(info.getTopicName().equals(this.topicName));
     }
 
-    /**
-     * Method: testGetEventGroupId(String eventId,String groupId)
-     */
-    @Test(expected = BrokerException.class)
-    public void testGetEventGroupId() throws Exception {
-        this.weEventClient.getEvent("not exist", groupId);
-    }
 
     /**
      * Method: getEvent(String eventId)
